Allow hostname to be passed in
Some of our nodes have hostnames determined at runtime and cannot be placed in the .env. We also can't update process.env.DISCOVERY_HOSTNAME, because the variable is captured [on import](https://github.com/dashersw/node-discover/blob/master/lib/network.js#L11) instead of on instantiation. This small change allows the hostname to be determined by the consumer of the library.